Collections Page Revamp
Okay so, tons of changes here.

Feel free to comment stuff and ask what my dumbass reasoning was. Chances are, there won't really be one.  I revamped a lot of functions to make them a little more basic with tons of error logging. Those can be removed when we are ready to release it.